requirements.txt
Django==3.0.7
djangorestframework==3.14.0
psycopg2-binary==2.8.5
djangorestframework-simplejwt==5.2.2
python-dotenv==1.0.0
pillow==9.4.0

